Project Setup
Added main Light Bulb and its animation using Gsap. It features a tuggable switch to turn it on and off.

Added co centric circles to the background around the light bulb.

Added isFocusMode context.

Background reacts to FocusMode context.

Added a svg to decorate the dashboard. Added logic for it to transition between colours too.