Settings Page v2
Fixed issue where focus mode would not change at all due to async errors with chrome local storage.

Fixed a lot of bugs with the orignal settings page but ended up revamping it completely. it is now inline with the main popup page in terms of design. It features the same background switch depending on the mode. The bulb on the right half of the screen is now working.